+++ src/components/form-checkbox.js
@@ -18,6 +18,10 @@ class FormCheckbox extends React.Component {
       cssClasses = cssClasses.concat(this.props.className.split(' '));
     }
 
+    if (this.props.disabled) {
+      cssClasses.push('is-disabled');
+    }
+
     return cssClasses;
   }
 
@@ -35,7 +39,7 @@ class FormCheckbox extends React.Component {
     }
 
     return (
-      <FormLabel inputType="checkbox" inputAlignment={this.props.inputAlignment}>
+      <FormLabel className={this.getCssClasses().join(' ')} inputType="checkbox" inputAlignment={this.props.inputAlignment}>
         <input
           className="u-hide"
           type="checkbox"